requirements.txt
wxPython==4.1.1
matplotlib==3.3.3
numpy==1.19.3
scipy==1.5.4
pandas==1.1.5
scikit-learn==0.23.2
seaborn==0.11.0
statsmodels==0.12.1
Pypubsub==4.0.3
